+++ website/api/comments/upvoteComment.ts
@@ -2,17 +2,17 @@ import axios from "axios";
 
 import apiBaseUrl from "../../utils/apiBaseUrl";
 
-export default function downvoteComment(commentId, parentItemId, callback) {
+export default function upvoteComment(commentId, parentItemId, callback) {
     axios
         .post(
-            apiBaseUrl + "/comments/downvote-comment",
+            apiBaseUrl + "/comments/upvote-comment",
             {
                 id: commentId,
                 parentItemId: parentItemId,
             },
             {
                 withCredentials: true,
-            },
+            }
         )
         .then(function (response) {
             callback(response.data);
