+++ website/components/comment.js
@@ -6,12 +6,12 @@ import renderPointsString from "../utils/renderPointsString";
 import renderCreatedTime from "../utils/renderCreatedTime";
 import truncateItemTitle from "../utils/truncateItemTitle";
 
-import addNewComment from "../api/comments/addNewComment.js";
-import upvoteComment from "../api/comments/upvoteComment.js";
-import downvoteComment from "../api/comments/downvoteComment.js";
-import unvoteComment from "../api/comments/unvoteComment.js";
-import favoriteComment from "../api/comments/favoriteComment.js";
-import unfavoriteComment from "../api/comments/unfavoriteComment.js";
+import addNewComment from "../api/comments/addNewComment";
+import upvoteComment from "../api/comments/upvoteComment";
+import downvoteComment from "../api/comments/downvoteComment";
+import unvoteComment from "../api/comments/unvoteComment";
+import favoriteComment from "../api/comments/favoriteComment";
+import unfavoriteComment from "../api/comments/unfavoriteComment";
 import killComment from "../api/moderation/killComment";
 import unkillComment from "../api/moderation/unkillComment";
