+++ website/components/commentSection.js
@@ -2,9 +2,9 @@ import { useState } from "react";
 import Link from "next/link";
 import Router from "next/router";
 
-import upvoteComment from "../api/comments/upvoteComment.js";
-import downvoteComment from "../api/comments/downvoteComment.js";
-import unvoteComment from "../api/comments/unvoteComment.js";
+import upvoteComment from "../api/comments/upvoteComment";
+import downvoteComment from "../api/comments/downvoteComment";
+import unvoteComment from "../api/comments/unvoteComment";
 
 import renderPointsString from "../utils/renderPointsString";
 import renderCreatedTime from "../utils/renderCreatedTime";
