+++ website/pages/reply.js
@@ -2,7 +2,7 @@ import AlternateHeader from "../components/AlternateHeader";
 import HeadMetadata from "../components/HeadMetadata";
 import Comment from "../components/comment.js";
 
-import getReplyPageData from "../api/comments/getReplyPageData.js";
+import getReplyPageData from "../api/comments/getReplyPageData";
 
 export default function Reply({ comment, authUserData, getDataError, notFoundError, goToString }) {
     return (
