+++ website/pages/upvoted.js
@@ -5,7 +5,7 @@ import ItemsList from "../components/itemsList.js";
 import CommentsList from "../components/commentsList.js";
 
 import getUserUpvotedItemsByPage from "../api/items/getUserUpvotedItemsByPage";
-import getUserUpvotedCommentsByPage from "../api/comments/getUserUpvotedCommentsByPage.js";
+import getUserUpvotedCommentsByPage from "../api/comments/getUserUpvotedCommentsByPage";
 
 export default function Upvoted({
     items,
